reducing failed tests
