--- clouddriver-openstack/src/main/groovy/com/netflix/spinnaker/clouddriver/openstack/security/LbaasVersion.groovy
@@ -1,9 +1,0 @@
-package com.netflix.spinnaker.clouddriver.openstack.security
-
-enum LbaasVersion {
-  V1, V2
-
-  String value() {
-    return toString().toLowerCase()
-  }
-}
