--- clouddriver-openstack/src/test/groovy/com/netflix/spinnaker/clouddriver/openstack/client/OpenstackIdentityV2ProviderSpec.groovy
@@ -1,54 +1,0 @@
-/*
- * Copyright 2016 Target, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.clouddriver.openstack.client
-
-import com.netflix.spinnaker.clouddriver.openstack.security.OpenstackNamedAccountCredentials
-import org.openstack4j.api.OSClient
-import spock.lang.Specification
-import spock.lang.Unroll
-
-
-@Unroll
-class OpenstackIdentityV2ProviderSpec extends Specification {
-
-  OSClient.OSClientV2 osClient
-  OpenstackNamedAccountCredentials namedAccountCredentials = Mock(OpenstackNamedAccountCredentials)
-
-  def "setup"() {
-    osClient = Mock(OSClient.OSClientV2)
-  }
-
-  def "test get regions - #testCase"() {
-    given:
-    OpenstackIdentityV2Provider provider = Spy(OpenstackIdentityV2Provider, constructorArgs:[namedAccountCredentials]) {
-      buildClient() >> { osClient }
-      getAllRegions() >> { expectedResult }
-    }
-
-    when:
-    List<String> result = provider.getAllRegions()
-
-    then:
-    result == expectedResult
-
-    where:
-    testCase | expectedResult
-    'empty'  | []
-    'null'   | null
-    'found'  | ['east', 'west']
-  }
-}
